pyproject.toml
[project]
name = "zindi-automation-submission-bot"
version = "1.0.0"
description = "Automated submission bot for Zindi competitions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rpaframework==28.6.3",
    "pre-commit>=2.17.0",
    "setuptools>=65.5.1",
    "pillow>=10.0.1",
    "pandas==2.1.4",
    "pydantic-yaml==1.2.0",
    "urllib3<=1.26.19",
    "selenium-stealth==1.0.6",
    "retry==0.9.2",
    "requests==2.32.2",
    "python-dateutil==2.8.2",
    "simplejson==3.19.2",
    "python-dotenv",
    "google-auth-oauthlib",
    "jira>=3.5.0",
    "office365-rest-python-client>=2.3.0",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=2.17.0",
]

[build-system]
requires = ["hatchling", "setuptools>=65.5.1"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["libraries", "Workflow"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=libraries",
    "--cov=Workflow",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["libraries", "Workflow"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

